Refactor ShareModal: separate controlled modal from auto-popup logic
- ShareModal is now a pure controlled component (open/onClose props only)
- ShareModalAutoPopup wraps it with timer, feature-flag, minimize, and dismiss logic for the results page
- Results.tsx uses ShareModalAutoPopup (identical behavior)
- Footer.tsx uses ShareModal directly (already correct)
