+++ src/Components/Results/Results.tsx
@@ -35,7 +35,7 @@ import { EnergyCalculatorRebateCategory } from '../EnergyCalculator/Results/reba
 import EnergyCalculatorRebatePage from '../EnergyCalculator/Results/RebatePage';
 import { usePageTitle } from '../Common/usePageTitle';
 import { NPSWidget } from '../NPS';
-import ShareModal from './ShareModal/ShareModal';
+import ShareModalAutoPopup from './ShareModal/ShareModalAutoPopup';
 
 type WrapperResultsContext = {
   programs: Program[];
@@ -263,7 +263,7 @@ const Results = ({ type }: ResultsProps) => {
               </Grid>
               {!noHelpButton && <HelpButton />}
               <NPSWidget uuid={uuid} />
-              <ShareModal />
+              <ShareModalAutoPopup />
             </div>
           </div>
         </main>
